moving more stuff into fitting package
