+++ ngmix/fitting/__init__.py
@@ -5,3 +5,10 @@ from .fitters import *
 
 from . import results
 from .results import *
+
+from . import galsim_fitters
+from .galsim_fitters import *
+from . import galsim_results
+from .galsim_results import *
+
+from . import leastsqbound
